+++ src/index.ts
@@ -9,6 +9,7 @@ import { regularWorkbenchColors } from "./ui/workbench_regular.js";
 import { contrastWorkbenchColors } from "./ui/workbench_contrast.js";
 import { vimColoring } from "./token/vim.js";
 import { tmTheme } from "./token/bat.js";
+import { windowsTerminalColors } from "./token/windows-terminal.js";
 import { VSPackage, VSTheme } from "./manifest/package.js";
 
 /* eslint-disable-next-line  @typescript-eslint/no-explicit-any */
@@ -274,6 +275,17 @@ function batColorScheme() {
   });
 }
 
+function windowsTerminalColorScheme() {
+  const palette = new (colorPaletteFactory())("Night Coder");
+  fs.writeFile("./windows-terminal/night-coder.json", windowsTerminalColors(palette), (err) => {
+    if (err) {
+      throw err;
+    }
+    console.log(`Windows Terminal color scheme generated.`);
+  });
+}
+
 vscodeThemesWriter();
 vimColorScheme();
 batColorScheme();
+windowsTerminalColorScheme();