+++ src/token/windows-terminal.ts
@@ -1,0 +1,35 @@
+import { Palette } from "../common/colors.js";
+
+interface WindowsTerminalSetting {
+  [prop: string]: string;
+}
+
+export function windowsTerminalColors(palette: Palette): string {
+  const settings: WindowsTerminalSetting = {
+    name: "Night Coder",
+    cursorColor: palette.backgroundee.code,
+    selectionBackground: `${palette.backgroundee.code}`,
+    background: palette.background.code,
+    foreground: palette.backgroundc5.code,
+
+    black: palette.ansiBlack.code,
+    blue: palette.ansiBlue.code,
+    cyan: palette.ansiCyan.code,
+    green: palette.ansiGreen.code,
+    purple: palette.ansiMagenta.code,
+    red: palette.ansiRed.code,
+    white: palette.ansiWhite.code,
+    yellow: palette.ansiYellow.code,
+
+    brightBlack: palette.ansiBrightBlack.code,
+    brightBlue: palette.ansiBrightBlue.code,
+    brightCyan: palette.ansiBrightCyan.code,
+    brightGreen: palette.ansiBrightGreen.code,
+    brightPurple: palette.ansiBrightMagenta.code,
+    brightRed: palette.ansiBrightRed.code,
+    brightWhite: palette.ansiBrightWhite.code,
+    brightYellow: palette.ansiBrightYellow.code,
+  };
+
+  return JSON.stringify(settings);
+}